tidy up - created separate classes for tempmod and messages - changed roll to use DiceHelper - removed old code
