+++ src/main/java/ti4/commands/combat/CombatRoll.java
@@ -1,11 +1,10 @@
-package ti4.commands.special;
+package ti4.commands.combat;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.StringTokenizer;
 import java.util.stream.Collectors;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
@@ -14,12 +13,13 @@ import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import org.apache.commons.lang3.StringUtils;
 import ti4.commands.units.AddRemoveUnits;
-import ti4.generator.Mapper;
 import ti4.generator.TileHelper;
 import ti4.helpers.AliasHandler;
 import ti4.helpers.CombatHelper;
+import ti4.helpers.CombatMessageHelper;
 import ti4.helpers.CombatModHelper;
 import ti4.helpers.CombatRollType;
+import ti4.helpers.CombatTempModHelper;
 import ti4.helpers.Constants;
 import ti4.helpers.Emojis;
 import ti4.helpers.Helper;
@@ -29,27 +29,19 @@ import ti4.map.Player;
 import ti4.map.Tile;
 import ti4.map.UnitHolder;
 import ti4.message.MessageHelper;
-import ti4.model.CombatModifierModel;
 import ti4.model.NamedCombatModifierModel;
-import ti4.model.PlanetModel;
 import ti4.model.TileModel;
 import ti4.model.UnitModel;
 
-public class CombatRoll extends SpecialSubcommandData {
+public class CombatRoll extends CombatSubcommandData {
 
     public CombatRoll() {
         super(Constants.COMBAT_ROLL,
-                "*V2* *BETA* Combat rolls for units on tile. *Auto includes always on mods*");
+                "*V2* *BETA* Combat rolls for units on tile. *Auto includes modifiers*");
         addOptions(new OptionData(OptionType.STRING, Constants.TILE_NAME, "System/Tile name").setRequired(true)
                 .setAutoComplete(true));
-        addOptions(new OptionData(OptionType.STRING, Constants.COMBAT_MODIFIERS,
-                "+/- <unit type>. Eg -1 all, +2 mech. Temp ACs/PN/exhaust-tech mods")
-                .setRequired(false));
         addOptions(new OptionData(OptionType.STRING, Constants.PLANET,
                 "(optional) Planet to have combat on. Default is space combat.").setAutoComplete(true)
-                .setRequired(false));
-        addOptions(new OptionData(OptionType.STRING, Constants.COMBAT_EXTRA_ROLLS,
-                "comma list of <count> <unit> eg 2 fighter 1 dreadnought")
                 .setRequired(false));
         addOptions(new OptionData(OptionType.STRING, Constants.COMBAT_ROLL_TYPE,
                 "switch to afb/bombardment/spacecannonoffence")
@@ -63,9 +55,7 @@ public class CombatRoll extends SpecialSubcommandData {
         Game activeGame = getActiveGame();
 
         OptionMapping tileOption = event.getOption(Constants.TILE_NAME);
-        OptionMapping mods = event.getOption(Constants.COMBAT_MODIFIERS);
         OptionMapping planetOption = event.getOption(Constants.PLANET);
-        OptionMapping extraRollsOption = event.getOption(Constants.COMBAT_EXTRA_ROLLS);
         OptionMapping rollTypeOption = event.getOption(Constants.COMBAT_ROLL_TYPE);
 
         String userID = getUser().getId();
@@ -80,15 +70,6 @@ public class CombatRoll extends SpecialSubcommandData {
 
         if (tileOption == null) {
             return;
-        }
-        List<NamedCombatModifierModel> customMods = new ArrayList<>();
-        if (mods != null) {
-            customMods = parseCustomUnitMods(mods.getAsString());
-        }
-
-        HashMap<String, Integer> extraRollsParsed = new HashMap<>();
-        if (extraRollsOption != null) {
-            extraRollsParsed = parseUnits(extraRollsOption.getAsString());
         }
 
         String unitHolderName = Constants.SPACE;
@@ -121,12 +102,11 @@ public class CombatRoll extends SpecialSubcommandData {
             }
         }
 
-        secondHalfOfCombatRoll(player, activeGame, event, tile, unitHolderName, extraRollsParsed, customMods, rollType);
+        secondHalfOfCombatRoll(player, activeGame, event, tile, unitHolderName, rollType);
     }
 
     public void secondHalfOfCombatRoll(Player player, Game activeGame, GenericInteractionCreateEvent event, Tile tile,
             String unitHolderName,
-            HashMap<String, Integer> extraRollsParsed, List<NamedCombatModifierModel> customMods,
             CombatRollType rollType) {
         String sb = "";
         UnitHolder combatOnHolder = tile.getUnitHolders().get(unitHolderName);
@@ -181,52 +161,26 @@ public class CombatRoll extends SpecialSubcommandData {
                 rollType, activeGame);
 
         TileModel tileModel = TileHelper.getAllTiles().get(tile.getTileID());
-        List<NamedCombatModifierModel> autoMods = CombatModHelper.CalculateAutomaticMods(player, opponent,
+        List<NamedCombatModifierModel> modifiers = CombatModHelper.GetModifiers(player, opponent,
                 playerUnitsByQuantity, tileModel, activeGame, rollType, Constants.COMBAT_MODIFIERS);
 
-        List<NamedCombatModifierModel> autoExtraRolls = CombatModHelper.CalculateAutomaticMods(player, opponent,
+        List<NamedCombatModifierModel> extraRolls = CombatModHelper.GetModifiers(player, opponent,
                 playerUnitsByQuantity, tileModel, activeGame, rollType, Constants.COMBAT_EXTRA_ROLLS);
 
         // Check for temp mods
-        CombatModHelper.EnsureValidTempMods(player, tileModel, combatOnHolder);
-        CombatModHelper.InitializeNewTempMods(player, tileModel, combatOnHolder);
-        List<NamedCombatModifierModel> tempMods = new ArrayList<>(CombatModHelper.BuildCurrentRoundTempNamedModifiers(player, tileModel,
+        CombatTempModHelper.EnsureValidTempMods(player, tileModel, combatOnHolder);
+        CombatTempModHelper.InitializeNewTempMods(player, tileModel, combatOnHolder);
+        List<NamedCombatModifierModel> tempMods = new ArrayList<>(CombatTempModHelper.BuildCurrentRoundTempNamedModifiers(player, tileModel,
                 combatOnHolder, false, rollType));
         List<NamedCombatModifierModel> tempOpponentMods = new ArrayList<>();
         if(opponent != null){
-            tempOpponentMods = CombatModHelper.BuildCurrentRoundTempNamedModifiers(opponent, tileModel,
+            tempOpponentMods = CombatTempModHelper.BuildCurrentRoundTempNamedModifiers(opponent, tileModel,
                 combatOnHolder, true, rollType);
         }
-        
         tempMods.addAll(tempOpponentMods);
 
-        List<UnitModel> unitsInCombat = new ArrayList<>(playerUnitsByQuantity.keySet());
-        customMods = CombatModHelper.FilterRelevantMods(customMods, unitsInCombat, rollType);
-        autoMods = CombatModHelper.FilterRelevantMods(autoMods, unitsInCombat, rollType);
-
-        autoExtraRolls = CombatModHelper.FilterRelevantMods(autoExtraRolls, unitsInCombat, rollType);
-
-        String holderName = combatOnHolder.getName();
-        Planet holderPlanet = null;
-        if (combatOnHolder instanceof Planet) {
-            holderPlanet = (Planet) combatOnHolder;
-        }
-        if (holderPlanet != null) {
-            PlanetModel planetModel = Mapper.getPlanet(holderPlanet.getName());
-            holderName = planetModel.getName();
-        }
-
-        String combatTypeName = StringUtils.capitalize(holderName) + " combat";
-        if (rollType != CombatRollType.combatround) {
-            combatTypeName = rollType.getValue();
-            if (holderPlanet != null) {
-                combatTypeName += " on " + StringUtils.capitalize(holderName);
-            }
-        }
-        String message = String.format("**%s** rolls for %s on %s %s:  \n",
-                combatTypeName, player.getFactionEmoji(),
-                tile.getPosition(), Emojis.RollDice);
-        message += CombatHelper.RollForUnits(playerUnitsByQuantity, opponentUnitsByQuantity, autoExtraRolls, customMods, autoMods, tempMods, player,
+        String message = CombatMessageHelper.displayCombatSummary(player, tile, combatOnHolder, rollType);
+        message += CombatHelper.RollForUnits(playerUnitsByQuantity, opponentUnitsByQuantity, extraRolls, modifiers, tempMods, player,
                 opponent,
                 activeGame, rollType);
 
@@ -238,67 +192,5 @@ public class CombatRoll extends SpecialSubcommandData {
     @Override
     public void reply(SlashCommandInteractionEvent event) {
         super.reply(event);
-    }
-
-    private List<NamedCombatModifierModel> parseCustomUnitMods(String unitList) {
-        List<NamedCombatModifierModel> resultList = new ArrayList<>();
-        unitList = unitList.replace(", ", ",");
-        StringTokenizer unitListTokenizer = new StringTokenizer(unitList, ",");
-        while (unitListTokenizer.hasMoreTokens()) {
-            String unitListToken = unitListTokenizer.nextToken();
-            StringTokenizer unitInfoTokenizer = new StringTokenizer(unitListToken, " ");
-            int count = 1;
-            boolean numberIsSet = false;
-
-            String unit = "";
-            if (unitInfoTokenizer.hasMoreTokens()) {
-                String ifNumber = unitInfoTokenizer.nextToken();
-                try {
-                    count = Integer.parseInt(ifNumber);
-                    numberIsSet = true;
-                } catch (Exception e) {
-                    unit = AliasHandler.resolveUnit(ifNumber);
-                }
-            }
-            if (unitInfoTokenizer.hasMoreTokens() && numberIsSet) {
-                unit = AliasHandler.resolveUnit(unitInfoTokenizer.nextToken());
-            }
-
-            CombatModifierModel combatModifier = new CombatModifierModel();
-            combatModifier.setValue(count);
-            combatModifier.setScope(unit);
-            combatModifier.setPersistenceType("CUSTOM");
-            resultList.add(new NamedCombatModifierModel(combatModifier, ""));
-        }
-        return resultList;
-    }
-
-    private HashMap<String, Integer> parseUnits(String unitList) {
-        HashMap<String, Integer> resultList = new HashMap<>();
-        unitList = unitList.replace(", ", ",");
-        StringTokenizer unitListTokenizer = new StringTokenizer(unitList, ",");
-        while (unitListTokenizer.hasMoreTokens()) {
-            String unitListToken = unitListTokenizer.nextToken();
-            StringTokenizer unitInfoTokenizer = new StringTokenizer(unitListToken, " ");
-            int count = 1;
-            boolean numberIsSet = false;
-
-            String unit = "";
-            if (unitInfoTokenizer.hasMoreTokens()) {
-                String ifNumber = unitInfoTokenizer.nextToken();
-                try {
-                    count = Integer.parseInt(ifNumber);
-                    numberIsSet = true;
-                } catch (Exception e) {
-                    unit = AliasHandler.resolveUnit(ifNumber);
-                }
-            }
-            if (unitInfoTokenizer.hasMoreTokens() && numberIsSet) {
-                unit = AliasHandler.resolveUnit(unitInfoTokenizer.nextToken());
-            }
-
-            resultList.put(unit, count);
-        }
-        return resultList;
     }
 }
