+++ src/main/java/ti4/helpers/ButtonHelper.java
@@ -54,7 +54,7 @@ import ti4.commands.leaders.UnlockLeader;
 import ti4.commands.planet.PlanetAdd;
 import ti4.commands.player.SendDebt;
 import ti4.commands.player.Setup;
-import ti4.commands.special.CombatRoll;
+import ti4.commands.combat.CombatRoll;
 import ti4.commands.special.StellarConverter;
 import ti4.commands.status.Cleanup;
 import ti4.commands.status.ListTurnOrder;
@@ -3341,7 +3341,7 @@ public class ButtonHelper {
                 }
             }
         }
-        new CombatRoll().secondHalfOfCombatRoll(player, activeGame, event, activeGame.getTileByPosition(pos), unitHolderName, new HashMap<>(), new ArrayList<>(), rollType);
+        new CombatRoll().secondHalfOfCombatRoll(player, activeGame, event, activeGame.getTileByPosition(pos), unitHolderName, rollType);
         if(buttonID.contains("bombardment") && activeGame.getLaws().containsKey("conventions")){
             MessageHelper.sendMessageToChannel(event.getMessageChannel(), "This is a reminder that conventions of war is in play, so bombardment of cultural planets is illegal. Ignore this message if not relevant");
         }
